Remove 'transition: transform' from mobile nav due to chrome on iOS bug
WebKit Bug 228333 causes chrome on iOS to not properly use 'transition: transform' in CSS. When a website with that property is opened, it transitions normally.
However, when a new tab is opened and you go back to the original tab, the the transition: transform animation becomes slightly delayed. And if you keep switching tabs
then going back to the original tab, the issue worsens and the animation delays longer. The problem only happens on chrome on iOS, not safari.